+++ pkg/platform/aws/aws-ec2/aws-ec2-cloudlet.go
@@ -301,7 +301,7 @@ func (a *AwsEc2Platform) GetCloudletInfraResourcesInfo(ctx context.Context) ([]e
 	return []edgeproto.InfraResource{}, nil
 }
 
-func (a *AwsEc2Platform) GetClusterAdditionalResources(ctx context.Context, cloudlet *edgeproto.Cloudlet, vmResources []edgeproto.VMResource, infraResMap map[string]edgeproto.InfraResource) map[string]edgeproto.InfraResource {
+func (a *AwsEc2Platform) GetClusterAdditionalResources(ctx context.Context, cloudlet *edgeproto.Cloudlet, vmResources []edgeproto.VMResource) map[string]edgeproto.InfraResource {
 	resInfo := make(map[string]edgeproto.InfraResource)
 	return resInfo
 }
