+++ pkg/platform/aws/aws-eks/aws-eks.go
@@ -288,7 +288,7 @@ func getAwsEksResources(ctx context.Context, cloudlet *edgeproto.Cloudlet, resou
 }
 
 // called by controller, make sure it doesn't make any calls to infra API
-func (a *AwsEksPlatform) GetClusterAdditionalResources(ctx context.Context, cloudlet *edgeproto.Cloudlet, vmResources []edgeproto.VMResource, infraResMap map[string]edgeproto.InfraResource) map[string]edgeproto.InfraResource {
+func (a *AwsEksPlatform) GetClusterAdditionalResources(ctx context.Context, cloudlet *edgeproto.Cloudlet, vmResources []edgeproto.VMResource) map[string]edgeproto.InfraResource {
 	log.SpanLog(ctx, log.DebugLevelApi, "AwsEks GetClusterAdditionalResources", "cloudletKey", cloudlet.Key)
 	// resource name -> resource units
 	cloudletRes := map[string]string{
@@ -300,9 +300,6 @@ func (a *AwsEksPlatform) GetClusterAdditionalResources(ctx context.Context, clou
 	resInfo := make(map[string]edgeproto.InfraResource)
 	for resName, resUnits := range cloudletRes {
 		resMax := uint64(0)
-		if infraRes, ok := infraResMap[resName]; ok {
-			resMax = infraRes.InfraMaxValue
-		}
 		resInfo[resName] = edgeproto.InfraResource{
 			Name:          resName,
 			InfraMaxValue: resMax,
