+++ pkg/platform/azure/azure-aks.go
@@ -158,7 +158,7 @@ func (a *AzurePlatform) GetCloudletInfraResourcesInfo(ctx context.Context) ([]ed
 }
 
 // called by controller, make sure it doesn't make any calls to infra API
-func (a *AzurePlatform) GetClusterAdditionalResources(ctx context.Context, cloudlet *edgeproto.Cloudlet, vmResources []edgeproto.VMResource, infraResMap map[string]edgeproto.InfraResource) map[string]edgeproto.InfraResource {
+func (a *AzurePlatform) GetClusterAdditionalResources(ctx context.Context, cloudlet *edgeproto.Cloudlet, vmResources []edgeproto.VMResource) map[string]edgeproto.InfraResource {
 	return nil
 }
