+++ pkg/platform/k8s-operator/k8sop.go
@@ -123,7 +123,7 @@ func (k *K8sOperator) GetCloudletInfraResourcesInfo(ctx context.Context) ([]edge
 }
 
 // TODO
-func (k *K8sOperator) GetClusterAdditionalResources(ctx context.Context, cloudlet *edgeproto.Cloudlet, vmResources []edgeproto.VMResource, _ map[string]edgeproto.InfraResource) map[string]edgeproto.InfraResource {
+func (k *K8sOperator) GetClusterAdditionalResources(ctx context.Context, cloudlet *edgeproto.Cloudlet, vmResources []edgeproto.VMResource) map[string]edgeproto.InfraResource {
 	resInfo := make(map[string]edgeproto.InfraResource)
 	return resInfo
 }
