+++ pkg/platform/vmpool/vmpool-cloudlet.go
@@ -55,7 +55,7 @@ func (o *VMPoolPlatform) GetCloudletInfraResourcesInfo(ctx context.Context) ([]e
 	return []edgeproto.InfraResource{}, nil
 }
 
-func (o *VMPoolPlatform) GetClusterAdditionalResources(ctx context.Context, cloudlet *edgeproto.Cloudlet, vmResources []edgeproto.VMResource, infraResMap map[string]edgeproto.InfraResource) map[string]edgeproto.InfraResource {
+func (o *VMPoolPlatform) GetClusterAdditionalResources(ctx context.Context, cloudlet *edgeproto.Cloudlet, vmResources []edgeproto.VMResource) map[string]edgeproto.InfraResource {
 	resInfo := make(map[string]edgeproto.InfraResource)
 	return resInfo
 }
