+++ pkg/platform/vsphere/vsphere-cloudlet.go
@@ -134,7 +134,7 @@ func (s *VSpherePlatform) GetCloudletInfraResourcesInfo(ctx context.Context) ([]
 	return []edgeproto.InfraResource{}, nil
 }
 
-func (s *VSpherePlatform) GetClusterAdditionalResources(ctx context.Context, cloudlet *edgeproto.Cloudlet, vmResources []edgeproto.VMResource, infraResMap map[string]edgeproto.InfraResource) map[string]edgeproto.InfraResource {
+func (s *VSpherePlatform) GetClusterAdditionalResources(ctx context.Context, cloudlet *edgeproto.Cloudlet, vmResources []edgeproto.VMResource) map[string]edgeproto.InfraResource {
 	resInfo := make(map[string]edgeproto.InfraResource)
 	return resInfo
 }
